replace everything with a single history path
