+++ weave/ops_domain/run_ops.py
@@ -54,11 +54,8 @@ from . import wb_util
 from .. import engine_trace
 from .run_history import history_op_common
 
-
 # Important to re-export ops
-from .run_history import run_history_v1_legacy_ops
-from .run_history import run_history_v2_parquet_media
-from .run_history import run_history_v3_parquet_stream_optimized
+from .run_history import run_history_ops
 
 tracer = engine_trace.tracer()
