+++ weave/tests/test_panel_time_series.py
@@ -11,7 +11,7 @@ def test_panel_timeseries(user_by_api_key_in_env):
         run.log({"val": i, "cat": str(i % 2)})
     run.finish()
 
-    history_node = weave.ops.project(run.entity, run.project).run(run.id).history2()
+    history_node = weave.ops.project(run.entity, run.project).run(run.id).history()
     panel = TimeSeries(history_node)
     init_config_node = panel.initialize()
     init_config = weave.use(init_config_node)
